+++ tests/test_component.py
@@ -1,4 +1,4 @@
-from vexipy.component import Component, Subcomponent
+from vexipy.component import Product, Subcomponent
 
 
 def test_update_subcomponent():
@@ -7,19 +7,19 @@ def test_update_subcomponent():
     assert s.hashes["md5"] == "d41d8cd98f00b204e9800998ecf8427e"
 
 
-def test_update_component():
-    c = Component()
+def test_update_product():
+    c = Product()
     c = c.update(hashes={"md5": "d41d8cd98f00b204e9800998ecf8427e"})
     assert c.hashes["md5"] == "d41d8cd98f00b204e9800998ecf8427e"
 
 
 def test_append_subcomponent():
-    c = Component()
+    c = Product()
     s = Subcomponent()
     assert c.append_subcomponents(s)
 
 
 def test_extend_subcomponent():
-    c = Component()
+    c = Product()
     s = [Subcomponent() for _ in range(3)]
     assert c.extend_subcomponents(s)