+++ vexipy/_util.py
@@ -1,7 +1,12 @@
-"""Miscellaneous helper functions"""
+"""Miscellaneous helper functions."""
 
 from datetime import datetime, timezone
 
 
 def utc_now() -> datetime:
+    """
+    Returns the current UTC datetime.
+
+    :return: Current datetime in UTC timezone.
+    """
     return datetime.now(timezone.utc)